RQ: 图片选择器分类

+++ app/src/main/java/com/zhyen/android/utils/ScreenUtils.java
@@ -97,7 +97,7 @@ public class ScreenUtils {
         return manager.getDefaultDisplay().getRotation();
     }
 
-    public static double getDensity(Context context) {
+    public static float getDensity(Context context) {
         float density = context.getResources().getDisplayMetrics().density;
         Log.d(TAG, "getDensity: " + density);
         return density;
